Move instruction block to separate file
