+++ src/block.h
@@ -1,0 +1,6 @@
+#include "pdc.h"
+#include "runtime.h"
+#include "astnode.h"
+
+void instructionBlockResolver(Runtime* r, ASTNode* self);
+ASTNode* createInstructionBlock(Instructions instructions);
